teacher_backend/requirements.txt
Django>=5.1,<6.0
openai>=1.40.0
openai-whisper>=20231117
librosa>=0.10
soundfile>=0.12
